Structure and generic parameters naming
