+++ Sources/Common/Item/ItemInteractor.swift
@@ -24,17 +24,3 @@ public class ItemInteractor<Item> {
   }
   
 }
-
-public extension ItemInteractor {
-  
-  func connectController<O, C: ItemController>(_ controller: C, dispatchOnMainThread: Bool = false, presenter: @escaping Presenter<Item, O>) where C.Item == O {
-    let sub = onItemChanged.subscribePast(with: controller) { controller, item in
-      controller.setItem(presenter(item))
-    }
-
-    if dispatchOnMainThread {
-      sub.onQueue(.main)
-    }
-  }
-  
-}
