+++ Sources/Number/FilterComparison+Controller.swift
@@ -9,6 +9,7 @@
 import Foundation
 
 extension NumberInteractor {
+  
   public func connectController<Controller: NumberController>(_ controller: Controller) where Controller.Item == Number {
 
     let computation = Computation(numeric: item) { [weak self] numeric in
@@ -25,4 +26,5 @@ extension NumberInteractor {
       controller.setItem(item)
     }
   }
+  
 }
